+++ src/Resources/app/administration/src/app/components/heptacom-admin-open-auth-user-confirm-login/index.js
@@ -25,6 +25,7 @@ export default {
             clients: [],
             waitingForConfirmation: false,
             confirmationClient: null,
+            popupsAreBlocked: null,
         }
     },
 
@@ -65,7 +66,9 @@ export default {
         },
 
         startAuthFlow(client) {
+            const that = this;
             this.waitingForConfirmation = true;
+            this.popupsAreBlocked = null;
             this.confirmationClient = client;
             localStorage.removeItem(confirmStateKey);
 
@@ -78,9 +81,33 @@ export default {
                         this.$t('heptacom-admin-open-auth-user-confirm-login.confirmWith', { 'clientName': client.name }),
                         `location=0,status=0,width=600,height=600, top=${top}, left=${left}`
                     );
-                    const windowLoop = window.setInterval(() => {
+                    let windowLoop = null;
+                    const windowPopupCheck = window.setTimeout(() => {
+                        if (!oauthWindow || oauthWindow.closed || typeof oauthWindow.closed === 'undefined') {
+                            that.popupsAreBlocked = true;
+                            that.waitingForConfirmation = false;
+
+                            if (windowLoop) {
+                                window.clearInterval(windowLoop);
+                            }
+                        }
+                    }, 1200); // must be longer than the loop, so the loop can run at least once
+
+                    try {
+                        oauthWindow.focus();
+                        that.popupsAreBlocked = false;
+                    } catch (e) {
+                        that.popupsAreBlocked = true;
+                        that.waitingForConfirmation = false;
+                        window.clearTimeout(windowPopupCheck);
+                        return;
+                    }
+
+                    windowLoop = window.setInterval(() => {
                         if (!this.waitingForConfirmation) {
                             window.clearInterval(windowLoop);
+                            window.clearTimeout(windowPopupCheck);
+                            this.popupsAreBlocked = false;
                             oauthWindow.close();
 
                             return;
@@ -88,6 +115,8 @@ export default {
 
                         if (oauthWindow.closed) {
                             window.clearInterval(windowLoop);
+                            window.clearTimeout(windowPopupCheck);
+                            this.popupsAreBlocked = false;
 
                             this.waitingForConfirmation = false;
                             const statePayload = localStorage.getItem(confirmStateKey);